# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET Saathi Financial Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SignupRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface PortfolioStock {
  symbol: string;
  name: string;
  quantity: number;
  avgPrice: number;
  currentPrice?: number;
  allocation?: number;
}

export type UserProfileRiskLevel =
  (typeof UserProfileRiskLevel)[keyof typeof UserProfileRiskLevel];

export const UserProfileRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UserProfile {
  id: number;
  userId: number;
  riskLevel: UserProfileRiskLevel;
  investmentGoal?: string;
  portfolio: PortfolioStock[];
  totalInvested?: number;
  updatedAt?: string;
}

export type UpdateProfileRequestRiskLevel =
  (typeof UpdateProfileRequestRiskLevel)[keyof typeof UpdateProfileRequestRiskLevel];

export const UpdateProfileRequestRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateProfileRequest {
  riskLevel?: UpdateProfileRequestRiskLevel;
  investmentGoal?: string;
  portfolio?: PortfolioStock[];
}

export interface NewsAnalysisRequest {
  newsText: string;
  userId?: number;
}

export type AffectedSectorImpact =
  (typeof AffectedSectorImpact)[keyof typeof AffectedSectorImpact];

export const AffectedSectorImpact = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export interface AffectedSector {
  name: string;
  impact: AffectedSectorImpact;
  magnitude: number;
}

export type NewsAnalysisResultSentiment =
  (typeof NewsAnalysisResultSentiment)[keyof typeof NewsAnalysisResultSentiment];

export const NewsAnalysisResultSentiment = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export interface NewsAnalysisResult {
  id?: number;
  keyEvent: string;
  eventType: string;
  affectedSectors: AffectedSector[];
  sentiment: NewsAnalysisResultSentiment;
  sentimentScore: number;
  marketImpact: string;
  summary: string;
  riskWarnings?: string[];
  confidence: number;
  createdAt?: string;
}

export interface ScenarioRequest {
  query: string;
  userId?: number;
}

export type ImpactedCompanyDirection =
  (typeof ImpactedCompanyDirection)[keyof typeof ImpactedCompanyDirection];

export const ImpactedCompanyDirection = {
  up: "up",
  down: "down",
  neutral: "neutral",
} as const;

export interface ImpactedCompany {
  name: string;
  symbol: string;
  impact: number;
  direction: ImpactedCompanyDirection;
  reason: string;
}

export type SectorImpactDirection =
  (typeof SectorImpactDirection)[keyof typeof SectorImpactDirection];

export const SectorImpactDirection = {
  up: "up",
  down: "down",
  neutral: "neutral",
} as const;

export interface SectorImpact {
  sector: string;
  change: number;
  direction: SectorImpactDirection;
}

export interface ScenarioResult {
  id?: number;
  scenario: string;
  niftyImpact: number;
  sectorImpacts: SectorImpact[];
  impactedCompanies: ImpactedCompany[];
  portfolioImpact?: number;
  actionableInsights: string[];
  explanation: string;
  confidence: number;
  timeframe?: string;
  createdAt?: string;
}

export interface DecisionRequest {
  stockSymbol: string;
  newsContext?: string;
  userId?: number;
}

export type DecisionResultRecommendation =
  (typeof DecisionResultRecommendation)[keyof typeof DecisionResultRecommendation];

export const DecisionResultRecommendation = {
  buy: "buy",
  hold: "hold",
  avoid: "avoid",
} as const;

export type DecisionResultRiskLevel =
  (typeof DecisionResultRiskLevel)[keyof typeof DecisionResultRiskLevel];

export const DecisionResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface DecisionResult {
  id?: number;
  stock: string;
  recommendation: DecisionResultRecommendation;
  confidence: number;
  riskLevel: DecisionResultRiskLevel;
  targetPrice?: number;
  stopLoss?: number;
  timeHorizon?: string;
  reasons: string[];
  explanation: string;
  warnings?: string[];
  createdAt?: string;
}

export interface SimulationRequest {
  stockSymbol: string;
  amount: number;
  years: number;
  userId?: number;
}

export interface SimulationDataPoint {
  year: number;
  investedValue: number;
  notInvestedValue: number;
}

export interface SimulationResult {
  id?: number;
  stock: string;
  initialAmount: number;
  projectedValue: number;
  notInvestedValue: number;
  annualReturn: number;
  totalReturn: number;
  dataPoints: SimulationDataPoint[];
  assumptions: string[];
  explanation: string;
  createdAt?: string;
}

export type HistoryItemType =
  (typeof HistoryItemType)[keyof typeof HistoryItemType];

export const HistoryItemType = {
  news: "news",
  scenario: "scenario",
  decision: "decision",
  simulation: "simulation",
} as const;

export type HistoryItemResult = { [key: string]: unknown };

export interface HistoryItem {
  id: number;
  type: HistoryItemType;
  query: string;
  result?: HistoryItemResult;
  createdAt: string;
}

export interface HistoryResponse {
  items: HistoryItem[];
}