# lib/api-zod/src/generated/types/newsAnalysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET Saathi Financial Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsAnalysisRequest {
  newsText: string;
  userId?: number;
}